third_party/blink/web_tests/fast/html/imports/encoding-expected.txt

CONSOLE WARNING: HTML Imports is deprecated and will be removed in M80, around February 2020. Please use ES modules instead. See https://www.chromestatus.com/features/5144752345317376 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Test that UTF-8 decoding is applied appropriately to imports

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Default encoding should be UTF-8...
PASS links[0].import.characterSet is "UTF-8"
And a meta tag should not override it...
PASS links[1].import.characterSet is "UTF-8"
PASS successfullyParsed is true

TEST COMPLETE
